test(atelier): split the backslash-escape guard tests into their own file
